regularized cov factorization
